# apispec_1.json
{
...
  "info": {
    "description": "powered by Flasgger",
    "termsOfService": "/tos",
    "title": "A swagger API",
    "version": "0.0.1"
  },
  "paths": {
...
    "/api/commentaire_petition/add": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "citoyen_id": {
                  "example": 1,
                  "type": "integer"
                },
                "description": {
                  "example": "Ceci est un commentaire.",
                  "type": "string"
                },
                "petition_id": {
                  "example": 1,
                  "type": "integer"
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Commentaire cr\u00e9\u00e9 avec succ\u00e8s"
          },
          "400": {
            "description": "Donn\u00e9es incompl\u00e8tes"
          },
          "500": {
            "description": "Erreur serveur"
          }
        },
        "summary": "Ajoute un nouveau commentaire \u00e0 une p\u00e9tition."
      }
    },
...
    "/api/commentaire_petition/delete/{commentaire_id}": {
      "delete": {
        "parameters": [
          {
            "example": 1,
            "in": "path",
            "name": "commentaire_id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "204": {
            "description": "Suppression r\u00e9ussie"
          },
          "404": {
            "description": "Commentaire non trouv\u00e9"
          },
          "500": {
            "description": "Erreur serveur"
          }
        },
        "summary": "Supprime un commentaire de la base de donn\u00e9es."
      }
    },
    "/api/commentaire_petition/update/{commentaire_id}": {
      "put": {
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "example": 1,
            "in": "path",
            "name": "commentaire_id",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "description": {
                  "example": "Nouveau contenu du commentaire.",
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Commentaire mis \u00e0 jour avec succ\u00e8s"
          },
          "400": {
            "description": "Donn\u00e9es incompl\u00e8tes"
          },
          "404": {
            "description": "Commentaire non trouv\u00e9"
          },
          "500": {
            "description": "Erreur serveur"
          }
        },
        "summary": "Met \u00e0 jour un commentaire existant."
      }
    },
    "/api/commentaire_petition/{citoyen_id}/citoyens": {
      "get": {
        "parameters": [
          {
            "example": 1,
            "in": "path",
            "name": "citoyen_id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des commentaires"
          }
        },
        "summary": "R\u00e9cup\u00e8re tous les commentaires d'un citoyen sp\u00e9cifique."
      }
    },
    "/api/commentaire_petition/{commentaire_id}": {
      "get": {
        "parameters": [
          {
            "example": 1,
            "in": "path",
            "name": "commentaire_id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "D\u00e9tails du commentaire"
          },
          "404": {
            "description": "Commentaire non trouv\u00e9"
          }
        },
        "summary": "R\u00e9cup\u00e8re un commentaire sp\u00e9cifique via son ID."
      }
    },
    "/api/commentaire_petition/{petition_id}/petitions": {
      "get": {
        "parameters": [
          {
            "example": 1,
            "in": "path",
            "name": "petition_id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des commentaires associ\u00e9s"
          }
        },
        "summary": "R\u00e9cup\u00e8re tous les commentaires li\u00e9s \u00e0 une p\u00e9tition sp\u00e9cifique."
      }
    },
...
    "/api/petition/add": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "cible": {
                  "example": "Gouvernement",
                  "type": "string"
                },
                "citoyen_id": {
                  "example": 1,
                  "type": "integer"
                },
                "description": {
                  "example": "P\u00e9tition pour am\u00e9liorer les infrastructures",
                  "type": "string"
                },
                "titre": {
                  "example": "Am\u00e9lioration des routes",
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "P\u00e9tition cr\u00e9\u00e9e avec succ\u00e8s"
          },
          "400": {
            "description": "Donn\u00e9es incompl\u00e8tes"
          },
          "500": {
            "description": "Erreur serveur"
          }
        },
        "summary": "Ajoute une nouvelle p\u00e9tition."
      }
    },
...
    "/api/petition/delete/{petition_id}": {
      "delete": {
        "parameters": [
          {
            "example": 1,
            "in": "path",
            "name": "petition_id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "204": {
            "description": "Suppression r\u00e9ussie"
          },
          "404": {
            "description": "P\u00e9tition non trouv\u00e9e"
          },
          "500": {
            "description": "Erreur serveur"
          }
        },
        "summary": "Supprime une p\u00e9tition de la base de donn\u00e9es."
      }
    },
    "/api/petition/update/{petition_id}": {
      "put": {
        "parameters": [
          {
            "example": 1,
            "in": "path",
            "name": "petition_id",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "description": {
                  "example": "Nouvelle description de la p\u00e9tition",
                  "type": "string"
                },
                "nb_signature": {
                  "example": 100,
                  "type": "integer"
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "P\u00e9tition mise \u00e0 jour avec succ\u00e8s"
          },
          "404": {
            "description": "P\u00e9tition non trouv\u00e9e"
          }
        },
        "summary": "Met \u00e0 jour les informations d'une p\u00e9tition existante."
      }
    },
    "/api/petition/{citoyen_id}/citoyens": {
      "get": {
        "parameters": [
          {
            "example": 1,
            "in": "path",
            "name": "citoyen_id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des p\u00e9titions"
          }
        },
        "summary": "R\u00e9cup\u00e8re toutes les p\u00e9titions d'un citoyen sp\u00e9cifique."
      }
    },
    "/api/petition/{petition_id}": {
      "get": {
        "parameters": [
          {
            "example": 1,
            "in": "path",
            "name": "petition_id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "D\u00e9tails de la p\u00e9tition"
          },
          "404": {
            "description": "P\u00e9tition non trouv\u00e9e"
          }
        },
        "summary": "R\u00e9cup\u00e8re une p\u00e9tition sp\u00e9cifique via son ID."
      }
    },
...
  },
  "swagger": "2.0"
}